backend/requirements.txt
playwright==1.44.0
fastapi==0.111.0
uvicorn==0.30.1
supabase==2.5.3
apscheduler==3.10.4
python-dotenv==1.0.1
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
resend==0.8.0
pydantic==2.7.4
